tool_generation_backend/pyproject.toml
[project]
name = "agent-browser-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.12.15",
    "fastapi>=0.116.2",
    "httpx>=0.28.1",
    "motor>=3.7.1",
    "openai>=1.0.0",
    "pydantic-settings>=2.10.1",
    "python-multipart>=0.0.20",
    "uvicorn[standard]>=0.35.0",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "mypy>=1.18.1",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "ruff>=0.13.0",
]
